Restructure Analytics styles
